requirements.txt
python-craigslist>=1.0.3
sqlalchemy
python-dateutil
ipython
slackclient
walkscore
